apps/salvo-matteini-bot/requirements.txt
numpy
pandas
keras
tensorflow
nltk
tweet-preprocessor